address review: move suggestLinks to shared, fix AGENTS.md prose
- suggestLinks extracted to src/renderer/shared/suggest-links.ts — it's an
  editor feature used in both notes and the event editor; EventEditorModal
  had a direct duplicate which is now removed
- notes/domain/link-resolution.ts retains only the notes-navigation helpers
  (resolveLinkById, resolveMarkdownHref)
- AGENTS.md: remove "live markdown editor" from description, fix component
  filenames to kebab-case, clarify hook/component conventions, restate
  the orchestrator ceiling in terms of "sequencer not logic owner" and the
  correct extraction pattern (pure computation → domain/, no setter-threading)
- CLAUDE.md: remove the test-stub heuristic sentence
- link-resolution.test.ts: clarify why event path strips the folder prefix
- scan-folder.test.ts: add test for paths containing spaces
