+++ src/renderer/timeline/event-editor/EventEditorModal.tsx
@@ -1,7 +1,7 @@
 import { useState, useEffect, useRef, useCallback, useMemo } from 'react';
 import type { EditorView } from '@codemirror/view';
 import { MarkdownEditor, FormatToolbar } from '../../shared/markdown-editor';
-import type { WikiLinkSuggestion } from '../../shared/markdown-editor';
+import { suggestLinks } from '../../shared/suggest-links';
 import { FooterPortal } from '../../components/footer-portal';
 import { openFromWikiLink, closeFromWikiLink } from '../../peek/stack';
 import { timelinePort, ConflictError } from '../data/ports';
@@ -64,17 +64,8 @@ export function EventEditorModal({
       .catch(() => {});
   }, [campaignPath]);
 
-  const suggestLinks = useCallback(
-    async (query: string): Promise<WikiLinkSuggestion[]> => {
-      const q = query.toLowerCase();
-      return linkIndex
-        .filter((e) => e.title.toLowerCase().includes(q) || e.id.toLowerCase().includes(q))
-        .map((e) =>
-          e.type === 'asset'
-            ? { id: '', label: e.title, detail: e.path, assetPath: e.path }
-            : { id: e.id, label: e.title, detail: e.path },
-        );
-    },
+  const suggestLinksForIndex = useCallback(
+    (query: string) => suggestLinks(linkIndex, query),
     [linkIndex],
   );
 
@@ -485,7 +476,7 @@ export function EventEditorModal({
                   onChange={(s) => updateBuffer({ body: s })}
                   viewRef={viewRef}
                   wikiLinks={{
-                    suggest: suggestLinks,
+                    suggest: suggestLinksForIndex,
                     onOpen: onOpenById,
                     onHover: openFromWikiLink,
                     onHoverEnd: closeFromWikiLink,
